[fix] Fix S3 Delete Error and Add Directory Option
